fix: properly fix tabs for production build
To get the frontend Javascript properly initialised, including in the
production build where files are stored wtih cache-busting hashes and so are
not able to be determined in advance, the only way I've found is to have it in
the .md files themselves. Slightly odd, but not the end of the world.